rename src to server
