+++ .dockerignore
@@ -2,7 +2,7 @@
 *
 
 # Show only the files we need
-!src/
+!server/
 !package.json
 !package-lock.json
 !tsconfig.json
