+++ Dockerfile.dev
@@ -19,11 +19,11 @@ RUN npm install -g npm && \
     npm install -g rimraf && \
     npm install 
 
-COPY ./src /app/src
+COPY ./server /app/server
 
 RUN npm run build
 
 EXPOSE 8000
-VOLUME ["/app/src"]
+VOLUME ["/app/server"]
 
 CMD ["npm", "start"]
